requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
psycopg2-binary==2.9.1
django-environ==0.8.1
djangorestframework>=3.12,<3.13